created new component, added template code
added html, css to job-actions-component, removed lines from job-result-component

removed html from job-result-component file

added new component in job-result template

renamed selector, removed unused objects

moved the new component to dashboard/job-result

passed Job object from job-result to job-actions-menu

removed ngOnInit(), ngOnDestroy() and polling stuff

renamed jobTableData prop to job

restored code for status icon

renamed jobTableData to job in job-actions-component.html

fixed alignment issues in component buttons

removed unused import

removed unused variables and imports

replaced jobID with job.id
