+++ src/app/pages/dashboard/job-result/job-actions-menu/job-actions-menu.component.html
@@ -1,0 +1,31 @@
+<!-- Download Sample btn -->
+<nb-icon
+  *ngIf="job.is_sample"
+  class="dl-btn"
+  nbTooltip="Download File"
+  icon="download-outline"
+  (click)="getJobSample()"
+></nb-icon>
+<!-- Raw JSON -->
+<nb-icon
+  class="dl-btn"
+  nbTooltip="View raw JSON"
+  icon="code-outline"
+  (click)="getJobRawJson()"
+></nb-icon>
+<!-- Kill Job Button -->
+<nb-icon
+  *ngIf="job.status === 'running'"
+  class="dl-btn"
+  status="warning"
+  nbTooltip="kill running analyzers and mark as 'killed'"
+  icon="slash"
+  (click)="killJob()"
+></nb-icon>
+<!-- Delete Job Button -->
+<nb-icon
+  status="danger"
+  nbTooltip="Delete Job"
+  icon="trash-2-outline"
+  (click)="deleteJob()"
+></nb-icon>
