+++ src/app/pages/dashboard/job-result/job-actions-menu/job-actions-menu.component.scss
@@ -1,0 +1,5 @@
+.dl-btn {
+  margin-right: 6.5px;
+  cursor: pointer;
+  color: inherit;
+}
