+++ src/app/pages/dashboard/job-result/job-result.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy, ViewChild } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { JobService } from '../../../@core/services/job.service';
 import { JobStatusIconRenderComponent } from '../../../@theme/components/smart-table/smart-table';
 import { Job } from '../../../@core/models/models';
@@ -8,7 +8,6 @@ import { Subscription } from 'rxjs';
 import { JsonEditorComponent, JsonEditorOptions } from 'ang-jsoneditor';
 import { trigger, transition, useAnimation } from '@angular/animations';
 import { flash } from 'ngx-animate';
-import { ToastService } from 'src/app/@core/services/toast.service';
 
 @Component({
   selector: 'intelowl-job-result',
@@ -89,9 +88,7 @@ export class JobResultComponent implements OnInit, OnDestroy {
 
   constructor(
     private readonly activateRoute: ActivatedRoute,
-    private readonly jobService: JobService,
-    private readonly router: Router,
-    private readonly toastr: ToastService
+    private readonly jobService: JobService
   ) {
     this.sub = this.activateRoute.params.subscribe(
       (res) => (this.jobId = res.jobId)
@@ -156,56 +153,6 @@ export class JobResultComponent implements OnInit, OnDestroy {
     }
     // finally assign it to our class' member variable
     this.jobTableData = res;
-  }
-
-  async getJobSample(): Promise<void> {
-    const url: string = await this.jobService.downloadJobSample(this.jobId);
-    window.open(url, 'rel=noopener,noreferrer');
-  }
-
-  async getJobRawJson(): Promise<void> {
-    const url: string = await this.jobService.downloadJobRawJson(this.jobId);
-    window.open(url, 'rel=noopener,noreferrer');
-  }
-
-  async deleteJob(): Promise<void> {
-    const sure = confirm('Are you sure?');
-    if (!sure) return;
-    const success = await this.jobService.deleteJobById(this.jobId);
-    if (success) {
-      this.ngOnDestroy();
-      this.toastr.showToast(
-        'Deleted successfully.',
-        `Job #${this.jobId}`,
-        'success'
-      );
-      setTimeout(() => this.router.navigate(['/']), 1000);
-    } else {
-      this.toastr.showToast(
-        'Could not be deleted. Reason: "Insufficient Permission".',
-        `Job #${this.jobId}`,
-        'error'
-      );
-    }
-  }
-
-  async killJob(): Promise<void> {
-    const sure = confirm('Are you sure?');
-    if (!sure) return;
-    const success = await this.jobService.killJobById(this.jobId);
-    if (success) {
-      this.toastr.showToast(
-        'Marked as "killed" successfully.',
-        `Job #${this.jobId}`,
-        'success'
-      );
-    } else {
-      this.toastr.showToast(
-        'Could not be "killed". Reason: "Insufficient Permission".',
-        `Job #${this.jobId}`,
-        'error'
-      );
-    }
   }
 
   // event emitted when user clicks on a row in table
